// src/bot/types/position.ts
export interface Position {
  id: string;
  address: string;
  tokenName: string;
  tokenSymbol: string;
  value: string; // Replace with the actual type for value
  profit: string; // Replace with the actual type for profit
  mcap: string; // Replace with the actual type for mcap
  tokenPrice: string; // Replace with the actual type for mcap
}